+++ lib/src/features/company_users/screen/company_users_list/company_users_list_screen_contents.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:irrigazione_iot/src/features/company_users/data/company_users_repository.dart';
-import 'package:irrigazione_iot/src/features/company_users/widgets/company_user_tile_widget.dart';
+import 'package:irrigazione_iot/src/features/company_users/widgets/company_user_list_tile.dart';
 import 'package:irrigazione_iot/src/features/company_users/widgets/empty_company_users_wid.dart';
 import 'package:irrigazione_iot/src/widgets/async_value_widget.dart';
 import 'package:irrigazione_iot/src/widgets/common_sliver_list_skeleton.dart';
